#56 Fixed unit test.
