+++ graphics/CMakeLists.txt
@@ -18,5 +18,5 @@ target_link_libraries(graphics ${GTK3_LIBRARIES})
 # Unit Tests
 set(MIN_UNIT_FILES ./tests/min_unit.cpp ./tests/min_unit.h ./tests/timer.cpp ./tests/timer.h)
 add_executable(coord_tests ./tests/coord_tests.cpp ${MIN_UNIT_FILES} ${SOURCE_FILES})
-add_executable(clipping_cs_tests tests/clipping_cs_tests.cpp ${MIN_UNIT_FILES} ${SOURCE_FILES})
+add_executable(region_tests tests/region_tests.cpp ${MIN_UNIT_FILES} ${SOURCE_FILES})
 add_executable(graphics_tests ./tests/graphics_tests.cpp ${MIN_UNIT_FILES} ${SOURCE_FILES})
